refactor(tests): centralize backend entity IDs in test-ids.ts
Extract TEST_BACKEND_USER_ID, OTHER_USER_ID, BLOCKED_USER_ID,
FOLLOW_REQUESTER_USER_ID, and TEST_PLAYER_ID into a dedicated
test-ids.ts file. Replaces magic numbers in mock-data fixtures and
keeps auth.fixture.ts focused on session/cookie concerns.
