+++ tests/fixtures/mock-data/me.ts
@@ -1,4 +1,5 @@
-import { TEST_BACKEND_USER_ID, TEST_PLAYER_ID, TEST_USER } from "../auth.fixture";
+import { TEST_USER } from "../auth.fixture";
+import { TEST_BACKEND_USER_ID, TEST_PLAYER_ID } from "../test-ids";
 
 export function mockMeResponse(overrides?: Record<string, unknown>) {
   return {
